facebook_grpc/requirements.txt
fbchat==1.9.6
grpcio-tools==1.27.2
